Change WebSocket upgrade methods
